requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.1
catalogue==1.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
cymem==2.0.11
docx2txt==0.8
en-core-web-sm==2.3.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
narwhals==1.25.2
nltk==3.9.1
numpy==1.23.5
packaging==24.2
pafy==0.5.5
pandas==2.2.3
pdfminer.six==20240706
pdfminer3==2018.12.3.0
pillow==11.1.0
plac==1.1.3
plotly==6.0.0
preshed==3.0.9
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.19.1
PyMySQL==1.1.1
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
spacy==2.3.5
srsly==1.0.7
streamlit==1.42.0
streamlit-option-menu==0.4.0
streamlit-tags==1.2.8
tenacity==9.0.0
thinc==7.4.6
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==0.10.1
watchdog==6.0.0
yt-dlp @ git+https://github.com/yt-dlp/yt-dlp.git@4ca8c44a073d5aa3a3e3112c35b2b23d6ce25ac6
